+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTitlePaneIcon.java
@@ -20,8 +20,6 @@ import java.awt.Dimension;
 import java.awt.Image;
 import java.util.ArrayList;
 import java.util.List;
-import javax.swing.Icon;
-import javax.swing.ImageIcon;
 import com.formdev.flatlaf.util.MultiResolutionImageSupport;
 import com.formdev.flatlaf.util.ScaledImageIcon;
 
@@ -31,40 +29,43 @@ import com.formdev.flatlaf.util.ScaledImageIcon;
 public class FlatTitlePaneIcon
 	extends ScaledImageIcon
 {
-	public static Icon create( List<Image> images, Dimension size ) {
-		// collect all images including multi-resolution variants
+	private final List<Image> images;
+
+	/**
+	 * @since 1.2
+	 */
+	public FlatTitlePaneIcon( List<Image> images, Dimension size ) {
+		super( null, size.width, size.height );
+		this.images = images;
+	}
+
+	@Override
+	protected Image getResolutionVariant( int destImageWidth, int destImageHeight ) {
+		// collect all images including multi-resolution variants for requested size
 		List<Image> allImages = new ArrayList<>();
 		for( Image image : images ) {
 			if( MultiResolutionImageSupport.isMultiResolutionImage( image ) )
-				allImages.addAll( MultiResolutionImageSupport.getResolutionVariants( image ) );
+				allImages.add( MultiResolutionImageSupport.getResolutionVariant( image, destImageWidth, destImageHeight ) );
 			else
 				allImages.add( image );
 		}
+
+		if( allImages.size() == 1 )
+			return allImages.get( 0 );
 
 		// sort images by size
 		allImages.sort( (image1, image2) -> {
 			return image1.getWidth( null ) - image2.getWidth( null );
 		} );
 
-		// create icon
-		return new FlatTitlePaneIcon( allImages, size );
-	}
-
-	private final List<Image> images;
-
-	private FlatTitlePaneIcon( List<Image> images, Dimension size ) {
-		super( new ImageIcon( images.get( 0 ) ), size.width, size.height );
-		this.images = images;
-	}
-
-	@Override
-	protected Image getResolutionVariant( int destImageWidth, int destImageHeight ) {
-		for( Image image : images ) {
+		// search for optimal image size
+		for( Image image : allImages ) {
 			if( destImageWidth <= image.getWidth( null ) &&
 				destImageHeight <= image.getHeight( null ) )
 			  return image;
 		}
 
-		return images.get( images.size() - 1 );
+		// use largest image
+		return allImages.get( allImages.size() - 1 );
 	}
 }